conf: reorganize/split code to idmap_utils.c
Move some idmaps-related functions from lxc/conf.c
to a new idmap_utils.c file.
